Update OWNERS file and use team email aliases.
Goal is to do a one-time synchronization of OWNERS file according to
information detailed in this page:
https://confluence.criteois.com/display/SE/Projects+Ownership

Once done OWNERS file will be the only reference for projects ownership, so
please ensure you keep them updated.

If you think this commit is inaccurate then maybe it has been generated with
outdated information. Please get in touch with me with an up-to-date list of
owners so I can fix this!

Change-Id: I9d0d605af92a0da76d078fcaf2b82bd1ba158981