JSON: Add the capability to read the raw serialized content of a node
The decoder is performing a check on the format while gathering the raw
content.

This feature can be useful when we don't know the schema of a JSON and
that we want to handle it as a raw string/dictionary; or when we want to
defer the deserialization of a sub part.